Restructured the code
